separate everything to abstartct and concrete
